Fixed the Fetching for all components, and moved serverIP. Issue#39
