--- serverIP.js
@@ -1,3 +1,0 @@
-export {serverIP}
-
-const serverIP = 'localhost:8080'; //remote server: 165.227.88.107
