+++ src/serverIP.js
@@ -1,0 +1,3 @@
+const serverIP = 'localhost:8080'; //remote server: 165.227.88.107
+
+export default {serverIP}
